Create Search component and searchTerm state
Create a search state within the main application and pass the state
fields as props to the Search component, for the sake of global scope.

Create the Search component in a seperate jsx file and let it read the
state fields as props.

Give the Search component an image element, and an input element that
listens for user keystrokes and dynamically updates the searchTerm
state.