Setup API call function & API error handling
Setup secure API key in seperate file within project.

Setup API_OPTIONS object which acts as a bridge of authentication and
ensuring proper data passing from API endpoint to program.

Setup API error handling for the API call function and display the error
in the console as well as on the screen.

Setup useEffect hook which will be the section of the code in which
we call the API call function.