Create practical webpage states: movieList & load
Create the loading state because the isLoading variable toggles between
true and false, and so that state needs to be remembered throughout
renders.

Add conditions for isLoading to be true: during the fetching of movies.
isLoading is false when displaying results: either error or movieList.

Create the movieList State because the webpage needs to remember what
set of movies (relative to the whole collection of movies) is currently
being displayed.