Create MovieCard component & add to movie section
Pass the movie prop to the MovieCard component so it can access the
specific pieces of data and present that data appropriately: the poster
paths for adding poster images, the vote average, the release date,
and the movie title.

For each movie inside, movieData instead of just displaying the movie
title, now display the MovieCard.