Implement Search
Create logic to seperate two different API calls. The first is the
discover API call which returns movies on a discover page, this is old
news. The second is the search API call which returns the results of a
search query provided by the user.

Pass a string prop "query" to the fetchMovies function and conditionally
enter the query into the search API call by altering the endpoint.

The API returns the search results.

Everytime the searchTerm value is updated (by the setter function within
the Search component), the useEffect hook recalls the fetchMovies
function. This is achieved by adding searchTerm to the dependency
array of useEffect.